+++ examples/mbed/main.cpp
@@ -6,7 +6,7 @@
  *
  * - keys must be 16 bytes
  * - EUIs must be 8 bytes
- * 
+ *
  * */
 const uint8_t app_key[] = MBED_CONF_APP_APP_KEY;
 const uint8_t nwk_key[] = MBED_CONF_APP_NWK_KEY;
@@ -14,21 +14,21 @@ const uint8_t dev_eui[] = MBED_CONF_APP_DEV_EUI;
 const uint8_t join_eui[] = MBED_CONF_APP_JOIN_EUI;
 
 SPI spi(D11, D12, D13);
-
-LDL::SX1272 radio(
-    spi,
-    D10,    // nss
-    A0,     // reset    
-    D2, D3  // DIO0, DIO1
-);
-
 LDL::DefaultSM sm(app_key, nwk_key);
 __attribute__ ((section (".noinit"))) LDL::DefaultStore store(dev_eui, join_eui);
-LDL::Device device(store, sm, radio);
 
 int main()
 {
     mbed_trace_init();
+
+    static LDL::SX1272 radio(
+        spi,
+        D10,    // nss
+        A0,     // reset
+        D2, D3  // DIO0, DIO1
+    );
+
+    static LDL::Device device(store, sm, radio);
 
     device.start(LDL_EU_863_870);
 
@@ -43,11 +43,11 @@ int main()
                 (void)device.unconfirmed(1, msg, strlen(msg));
             }
             else{
-            
+
                 (void)device.otaa();
             }
         }
     }
-    
+
     return 0;
 }