+++ wrappers/mbed/big_device.h
@@ -37,11 +37,17 @@ namespace LDL {
 
             MAC mac;
 
-            Thread event_thread;
+            Thread worker_thread;
 
             volatile bool done;
             Semaphore api;
             Mutex mutex;
+            Semaphore work;
+            LowPowerTimeout timeout;
+
+            void do_work();
+
+            void worker();
 
             void begin_api();
             void wait_until_api_done();
@@ -68,7 +74,6 @@ namespace LDL {
             Device(Store& store, SM& sm, Radio& radio);
 
             bool start(enum ldl_region region);
-            void stop();
 
             enum ldl_mac_status unconfirmed(uint8_t port, const void *data, uint8_t len, const struct ldl_mac_data_opts *opts = NULL);
             enum ldl_mac_status confirmed(uint8_t port, const void *data, uint8_t len, const struct ldl_mac_data_opts *opts = NULL);