+++ wrappers/mbed/mac.cpp
@@ -28,15 +28,13 @@ using namespace LDL;
 
 /* constructors *******************************************************/
 
-MAC::MAC(EventQueue& event, Store &store, SM &sm, Radio &radio) :
+MAC::MAC(Store &store, SM &sm, Radio &radio) :
     signal(A0, 0),
     radio(radio),
     sm(sm),
-    store(store),
-    event(event),
-    semaphore(1)
+    store(store)
 {
-    next_event_handler = 0;
+    timer.start();
 }
 
 /* protected static  **************************************************/
@@ -50,7 +48,7 @@ MAC::to_obj(void *self)
 uint32_t
 MAC::_ticks(void *app)
 {
-    return to_obj(app)->event.tick();
+    return (uint32_t)to_obj(app)->timer.elapsed_time().count();
 }
 
 uint32_t
@@ -211,49 +209,12 @@ MAC::app_handler(void *app, enum ldl_mac_response_type type, const union ldl_mac
 /* protected **********************************************************/
 
 void
-MAC::do_handle_radio_event(enum ldl_radio_event event, uint32_t ticks)
-{
-    LDL_MAC_radioEventWithTicks(&mac, event, ticks);
-    semaphore.release();
-}
-
-void
 MAC::handle_radio_event(enum ldl_radio_event event)
 {
-    if(semaphore.try_acquire()){
-
-        /* we don't assert on the return value
-         * since not having memory here will show up later as
-         * "chip error" when the interrupt never arrives.
-         *
-         * */
-        this->event.call(callback(this, &MAC::do_handle_radio_event), event, this->event.tick());
-    }
+    LDL_MAC_radioEvent(&mac, event);
 }
 
 /* public methods *****************************************************/
-
-void
-MAC::do_process()
-{
-    uint32_t next;
-
-    do{
-
-        LDL_MAC_process(&mac);
-        next = LDL_MAC_ticksUntilNextEvent(&mac);
-    }
-    while(next == 0U);
-
-    next = (next == UINT32_MAX) ? 60000UL : next;
-
-    event.cancel(next_event_handler);
-
-    next_event_handler = event.call_in(std::chrono::milliseconds(next), callback(this, &MAC::do_process));
-
-    /* LoRaWAN will grind to a halt if there is no memory to enqueue */
-    MBED_ASSERT(next_event_handler != 0);
-}
 
 bool
 MAC::start(enum ldl_region region)
@@ -273,12 +234,11 @@ MAC::start(enum ldl_region region)
     store.get_init_params(&store_params);
     session_size = store.get_session(&session, sizeof(session));
 
-    // clock source is the millisecond ticker in EventQueue
     arg.ticks = _ticks;
-    arg.tps = 1000UL;
-    arg.a = 0UL;
-    arg.b = 4UL;
-    arg.advance = 0UL;
+    arg.tps = MBED_CONF_LDL_TPS;
+    arg.a = MBED_CONF_LDL_XTAL_ERROR_A;
+    arg.b = MBED_CONF_LDL_XTAL_ERROR_B;
+    arg.advance = MBED_CONF_LDL_ADVANCE;
 
     arg.rand = _rand;
     arg.get_battery_level = _get_battery_level;
@@ -317,15 +277,7 @@ MAC::start(enum ldl_region region)
 
     run_state = ON;
 
-    do_process();
-
     return true;
-}
-
-void
-MAC::stop()
-{
-    run_state = OFF;
 }
 
 enum ldl_mac_status
@@ -421,5 +373,5 @@ MAC::ticks_until_next_event()
 void
 MAC::process()
 {
-    event.dispatch(0);
+    LDL_MAC_process(&mac);
 }